+++ CHANGELOG.md
@@ -20,6 +20,7 @@ the code already only really look there.
 - Tests can now use `value is not defined` order for negation (https://github.com/Keats/tera/issues/308)
 - Add `nth` filter to get the nth value in an array
 - You can now use glob patterns in Tera::new
+- `default` filter now works on Null values
 
 ## 0.11.20 (2018-11-14)